Refactor back to util class
