+++ src/nl/hannahsten/texifyidea/run/latex/LatexConfigurationFactory.kt
@@ -23,7 +23,7 @@ class LatexConfigurationFactory(type: ConfigurationType) : ConfigurationFactory(
             setDefaultPdfViewer()
             setDefaultOutputFormat()
             setSuggestedName()
-            setDefaultDistribution()
+            setDefaultDistribution(project)
         }
         is BibtexRunConfigurationType -> BibtexRunConfiguration(project, this, "BibTeX")
         is MakeindexRunConfigurationType -> MakeindexRunConfiguration(project, this, "Makeindex")
