+++ src/nl/hannahsten/texifyidea/util/files/LatexPackageLocationCache.kt
@@ -1,8 +1,8 @@
 package nl.hannahsten.texifyidea.util.files
 
 import com.intellij.openapi.project.Project
-import nl.hannahsten.texifyidea.settings.LatexSdk
-import nl.hannahsten.texifyidea.settings.LatexSdk.Companion.getLatexExecutableName
+import nl.hannahsten.texifyidea.settings.LatexSdkUtil
+import nl.hannahsten.texifyidea.settings.LatexSdkUtil.getLatexExecutableName
 import java.io.BufferedReader
 import java.io.IOException
 import java.io.InputStreamReader
@@ -29,7 +29,7 @@ object LatexPackageLocationCache {
     }
 
     private fun runKpsewhich(arg: String, project: Project): String? = try {
-        val command = if (LatexSdk.isMiktexAvailable) {
+        val command = if (LatexSdkUtil.isMiktexAvailable) {
             // Don't install the package if not present
             "miktex-kpsewhich --miktex-disable-installer $arg"
         }
