+++ test/nl/hannahsten/texifyidea/testutils/Common.kt
@@ -7,7 +7,7 @@ import io.mockk.mockkObject
 import io.mockk.mockkStatic
 import nl.hannahsten.texifyidea.run.compiler.LatexCompiler
 import nl.hannahsten.texifyidea.run.latex.LatexRunConfiguration
-import nl.hannahsten.texifyidea.settings.LatexSdk
+import nl.hannahsten.texifyidea.settings.LatexSdkUtil
 import nl.hannahsten.texifyidea.util.selectedRunConfig
 
 /**
@@ -37,7 +37,7 @@ fun setUnicodeSupport(project: Project, enabled: Boolean = true) {
     else {
         // Unicode is not supported on pdflatex on texlive <= 2017.
         every { project.selectedRunConfig()?.compiler } returns LatexCompiler.PDFLATEX
-        mockkObject(LatexSdk)
-        every { LatexSdk.texliveVersion } returns 2017
+        mockkObject(LatexSdkUtil)
+        every { LatexSdkUtil.texliveVersion } returns 2017
     }
 }
